1. Improving naming 2. Added PoolMessageDispatcher to dispatch IPublisherMessage by pools 3. Added PoolSnapshotActor to process the pool messages 4. Number of instances of PoolMessageDispatcher will increase per messagereceiver node and PoolSnapshotActor workload will be distributed throughout the cluster.

+++ src/MessagePublisher.Shared/Models/PoolInvestment.cs
@@ -4,11 +4,11 @@ namespace MessagePublisher.Shared.Models
 {
     public class PoolInvestment
     {
-        public long GameId { get; private set; }
+        public int GameId { get; private set; }
         public int PoolId { get; private set; }
         public List<CombinationInvestment> Combinations { get; private set; }
 
-        public PoolInvestment(long gameId, int poolId, List<CombinationInvestment> combinations)
+        public PoolInvestment(int gameId, int poolId, List<CombinationInvestment> combinations)
         {
             GameId = gameId;
             PoolId = poolId;
