+++ src/MessagePublisher.Shared/Models/PoolOddsChange.cs
@@ -4,11 +4,11 @@ namespace MessagePublisher.Shared.Models
 {
     public class PoolOddsChange
     {
-        public long GameId { get; private set; }
+        public int GameId { get; private set; }
         public int PoolId { get; private set; }
         public List<CombinationOddsChange> Combinations { get; private set; }
 
-        public PoolOddsChange(long gameId, int poolId, List<CombinationOddsChange> combinations)
+        public PoolOddsChange(int gameId, int poolId, List<CombinationOddsChange> combinations)
         {
             GameId = gameId;
             PoolId = poolId;
